fix循环依赖
